pyproject.toml
[project]
name = "ds-202501-health-econ-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hdbscan>=0.8.40",
    "ipykernel>=6.30.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
    "xlrd>=2.0.2",
]
